refactor(LoadingSkeleton): convert ContentCard to TypeScript
